remove tabs, they are not needed anymore because we have seperate links
